refactor: Adjust order of recommended extensions
Reorder the recommended extensions list to ensure logical grouping and consistency.
Removed "ms-azuretools.vscode-containers" from the "gitpodAlternativeRecommendations" array and added it to the main "recommendations" array.

This change maintains the list of suggested extensions for better readability and ensures that the extension is recommended in the primary list rather than as an alternative.

Signed-off-by: Banee Ishaque K <Baneeishaque@gmail.com>